Moved table_row and header widget back into widget crate.
