+++ crates/core/Cargo.toml
@@ -36,6 +36,7 @@ grin_wallet_controller = { git = "https://github.com/mimblewimble/grin-wallet", 
 grin_wallet_api = { git = "https://github.com/mimblewimble/grin-wallet", branch = "master" }
 grin_wallet_impls = { git = "https://github.com/mimblewimble/grin-wallet", branch = "master" }
 grin_wallet_libwallet = { git = "https://github.com/mimblewimble/grin-wallet", branch = "master" }
+grin-gui-widgets = { path = "../widgets" }
 
 
 ############ Local testing ################
