+++ crates/core/src/lib.rs
@@ -11,7 +11,6 @@ pub mod error;
 pub mod fs;
 pub mod theme;
 pub mod network;
-pub mod widgets;
 #[cfg(feature = "gui")]
 pub mod utility;
