+++ crates/core/src/theme/table_row.rs
@@ -1,4 +1,4 @@
-use crate::widgets::style::table_row::{StyleSheet, Appearance};
+use grin_gui_widgets::style::table_row::{StyleSheet, Appearance};
 use iced::{Background, Color};
 use super::Theme;
