--- crates/core/src/widgets/mod.rs
@@ -1,7 +1,0 @@
-pub mod style;
-pub mod widget;
-//pub mod renderer;
-
-pub use widget::header;
-pub use widget::table_row;
-
