--- crates/core/src/widgets/widget/mod.rs
@@ -1,2 +1,0 @@
-pub mod header;
-pub mod table_row;
