+++ crates/widgets/src/widget/table_row.rs
@@ -1,21 +1,18 @@
 #![allow(clippy::type_complexity)]
-
-pub use crate::style::table_row::{Style, StyleSheet};
-// use iced::Theme;
+use crate::style::table_row::StyleSheet;
+use iced::{Background, Color};
 use iced_native::{
     event, layout, mouse, overlay, renderer, widget, widget::Tree, Alignment, Clipboard, Element,
     Event, Layout, Length, Padding, Point, Rectangle, Shell, Widget,
 };
-use std::marker::PhantomData;
 
 #[allow(missing_debug_implementations)]
-pub struct TableRow<'a, Message, Renderer, Theme, Style>
+pub struct TableRow<'a, Message, Renderer>
 where
-    Renderer: 'a + self::Renderer<Theme, Style>,
+    Renderer: 'a + iced_native::Renderer,
+    Renderer::Theme: StyleSheet,
     Message: 'a,
 {
-    // we need to reference the genereic theme here so rust doesn't compain about Theme not being used
-    unused_arg: PhantomData<Theme>,
     padding: Padding,
     width: Length,
     height: Length,
@@ -24,14 +21,15 @@ where
     inner_row_height: u32,
     horizontal_alignment: Alignment,
     vertical_alignment: Alignment,
-    style_sheet: Box<dyn StyleSheet<Style = Style> + 'a>,
+    style: <Renderer::Theme as StyleSheet>::Style,
     content: Element<'a, Message, Renderer>,
     on_press: Option<Box<dyn Fn(Event) -> Message + 'a>>,
 }
 
-impl<'a, Message, Renderer, Theme, Style> TableRow<'a, Message, Renderer, Theme, Style>
+impl<'a, Message, Renderer> TableRow<'a, Message, Renderer>
 where
-    Renderer: 'a + self::Renderer<Theme, Style>,
+    Renderer: 'a + iced_native::Renderer,
+    Renderer::Theme: StyleSheet,
     Message: 'a,
 {
     /// Creates an empty [`TableRow`].
@@ -40,7 +38,6 @@ where
         T: Into<Element<'a, Message, Renderer>>,
     {
         TableRow {
-            unused_arg: PhantomData,
             padding: Padding::ZERO,
             width: Length::Shrink,
             height: Length::Shrink,
@@ -49,13 +46,18 @@ where
             inner_row_height: u32::MAX,
             horizontal_alignment: Alignment::Start,
             vertical_alignment: Alignment::Start,
-            style_sheet: Default::default(),
+            style:  Default::default(),
             content: content.into(),
             on_press: None,
         }
     }
-    pub fn style(mut self, style: impl Into<Box<dyn StyleSheet<Style = Style> + 'a>>) -> Self {
-        self.style_sheet = style.into();
+
+    /// Sets the style of the [`TableRow`].
+    pub fn style(
+        mut self,
+        style: impl Into<<Renderer::Theme as StyleSheet>::Style>,
+    ) -> Self {
+        self.style = style.into();
         self
     }
 
@@ -128,11 +130,11 @@ where
     }
 }
 
-impl<'a, Message, Renderer, Theme, Style> Widget<Message, Renderer>
-    for TableRow<'a, Message, Renderer, Theme, Style>
+impl<'a, Message, Renderer> Widget<Message, Renderer>
+    for TableRow<'a, Message, Renderer>
 where
-    Renderer: 'a + self::Renderer<Theme, Style>,
-    Renderer::Theme: iced::widget::container::StyleSheet + iced::widget::text::StyleSheet,
+    Renderer: 'a + iced_native::Renderer,
+    Renderer::Theme: StyleSheet,
     Message: 'a,
 {
     fn width(&self) -> Length {
@@ -168,30 +170,57 @@ where
         _tree: &Tree,
         renderer: &mut Renderer,
         theme: &Renderer::Theme,
-        _style: &renderer::Style,
+        style: &renderer::Style,
         layout: Layout<'_>,
         cursor_position: Point,
         viewport: &Rectangle,
     ) {
-        let mut bounds = layout.bounds();
+        let bounds = layout.bounds();
+        let mut custom_bounds = layout.bounds();
         let tree = Tree::new(&self.content); 
 
         // inner_row_height set?
         if self.inner_row_height != u32::MAX {
-            bounds.height = self.inner_row_height as f32;
+            custom_bounds.height = self.inner_row_height as f32;
         }
 
-        self::Renderer::draw(
-            renderer,
+        let is_mouse_over = custom_bounds.contains(cursor_position);
+        let content_layout = layout.children().next().unwrap();
+
+        let appearance = if is_mouse_over {
+            theme.hovered(&self.style)
+        } else {
+            theme.appearance(&self.style)
+        };
+
+        let background = iced_native::renderer::Quad {
+            bounds: Rectangle {
+                x: bounds.x + appearance.offset_left as f32,
+                y: bounds.y,
+                width: bounds.width - appearance.offset_right as f32,
+                height: custom_bounds.height,
+            },
+            border_radius: appearance.border_radius,
+            border_width: appearance.border_width,
+            border_color: appearance.border_color,
+        };
+
+        renderer.fill_quad(
+            background.into(),
+            appearance
+                .background
+                .unwrap_or(Background::Color(Color::TRANSPARENT)),
+        );
+
+        self.content.as_widget().draw(
             &tree,
-            layout,
+            renderer,
             theme,
+            style,
+            content_layout,
             cursor_position,
-            self.style_sheet.as_ref(),
-            &self.content,
             viewport,
-            &bounds,
-        )
+        );
     }
 
     fn mouse_interaction(
@@ -202,7 +231,15 @@ where
         viewport: &Rectangle,
         renderer: &Renderer,
     ) -> mouse::Interaction {
-        self::Renderer::mouse_interaction(renderer, layout, cursor_position, viewport)
+        
+        let bounds = layout.bounds();
+        let is_mouse_over = bounds.contains(cursor_position);
+
+        if is_mouse_over {
+            mouse::Interaction::Pointer
+        } else {
+            mouse::Interaction::default()
+        }
     }
 
     /*fn hash_layout(&self, state: &mut Hasher) {
@@ -275,40 +312,15 @@ where
     }
 }
 
-//use grin_gui_core::theme::Theme as Custom;
-//pub trait Renderer: iced_native::Renderer<Theme = iced_native::Theme> {
-pub trait Renderer<T, S>: iced_native::Renderer<Theme = T> {
-    type Style: Default;
 
-    #[allow(clippy::too_many_arguments)]
-    fn draw<Message>(
-        &mut self,
-        tree: &Tree,
-        layout: Layout<'_>,
-        theme: &T,
-        cursor_position: Point,
-        style_sheet: &dyn StyleSheet<Style = S>,
-        content: &Element<'_, Message, Self>,
-        viewport: &Rectangle,
-        custom_bounds: &Rectangle,
-    );
-
-    fn mouse_interaction(
-        &self,
-        layout: Layout<'_>,
-        cursor_position: Point,
-        viewport: &Rectangle,
-    ) -> mouse::Interaction;
-}
-
-impl<'a, Message, Renderer, Theme: 'a> From<TableRow<'a, Message, Renderer, Theme, Style>>
+impl<'a, Message, Renderer> From<TableRow<'a, Message, Renderer>>
     for Element<'a, Message, Renderer>
 where
-    Renderer: 'a + self::Renderer<Theme, Style>,
+    Renderer: 'a + iced_native::Renderer,
     Renderer::Theme: StyleSheet + widget::container::StyleSheet + widget::text::StyleSheet,
     Message: 'a,
 {
-    fn from(table_row: TableRow<'a, Message, Renderer, Theme, Style>) -> Element<'a, Message, Renderer> {
+    fn from(table_row: TableRow<'a, Message, Renderer>) -> Element<'a, Message, Renderer> {
         Element::new(table_row)
     }
 }
